🏗Multi-doc manager refactor (#25629)
* Separates multidoc-manager from the runtime

* Export installAmpdocServices

* Implements dynamic adoption of shadow mode

* Type check

* Fix type checking

* Revert dynamic adoption

* Export css dependency from multidoc-manager

* Removed dynamic asdoption of shadow mode

* Excludes inabox and video-iframe-integration from requiring Multidoc dependencies

* Type check fix
